backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
supabase
python-dotenv
gotrue
fastapi[all]